% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spCoordinateTrans.R
\name{addressToCoord}
\alias{addressToCoord}
\title{Address to Coordinates (WGS84)}
\usage{
addressToCoord(address, sleep_time = 0.2)
}
\arguments{
\item{address}{a character address}
}
\description{
Google maps 를 이용하여 주소정보를 대표좌표로 변환합니다.
}
\examples{
addressToCoord("강남구청역")

library(leaflet)
library(dplyr)

pd <- data.frame(add = c("강남구청역")) \%>\%
  mutate(long = addressToCoord(add)[1], lat = addressToCoord(add)[2])

leaflet(data = pd) \%>\% addTiles() \%>\%
  addMarkers(~long, ~lat, popup = ~as.character(add), label = ~as.character(add))
}
