% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv2rda.R
\name{csv2rda}
\alias{csv2rda}
\title{Write rda about imported tables}
\usage{
csv2rda(pattern = NULL, delete = F, format = c("csv", "txt"), ...)
}
\arguments{
\item{path}{a character}

\item{method}{choice machine learning algorithm method}

\item{fitImage}{a character. Input filename for chaching fit Object}
}
\value{
\code{rda} 포맷에서 객체의 이름은 \code{rawdata} 로 지정되어 저장됩니다.
}
\description{
워킹디렉토리에 존재하는 테이블 파일들을 import 한 후 rda 형태로 모두 변환합니다. 기호에 따라 변환되고 난 후 csv 파일을 지울 수 있습니다. 본 함수는 \code{readr::read_csv()} 함수에 의존성을 가지고 있습니다.
}
