% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{fitSummary}
\alias{fitSummary}
\title{Fitting summary}
\usage{
fitSummary(fitObject, testset = NULL, class = fitObject$class,
  type = c("cla", "reg"), ...)
}
\arguments{
\item{fitObject}{a "ml" class object}

\item{testset}{if new testset}

\item{method}{choice machine learning algorithm method}

\item{fitImage}{a character. Input filename for chaching fit Object}
}
\value{
\code{type} if "cla", confusion matrix. else "reg" numeric of MSE
}
\description{
\code{ml} class 인 학습결과 및 테스트셋을 이용하여 Classification 일 경우 혼돈메트릭스, Regression 일 경우 MSE 가 반환되며 \code{type} 인자를 통해 Classification 인지 Regression 인지 선택합니다.
}
\examples{
fit <- ml(iris, "Species")
fitSummary(fit, type = "cla")
}
