% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{ml}
\alias{ml}
\title{Machine learning framework using caret package}
\usage{
ml(data, class, method = "rpart", cvK = 2, cvRepeat = 2,
  partitionRate = 0.7, fitImage = NULL, ...)
}
\arguments{
\item{data}{a data.frame.}

\item{class}{a character. Input the class Y variable names}

\item{method}{choice machine learning algorithm method}

\item{fitImage}{a character. Input filename for chaching fit Object}
}
\description{
Machine learning framework using caret package
}
